+++ src/Language/Dickinson/Check/Duplicate.hs
@@ -47,3 +47,4 @@ checkExprDuplicates (Match _ e brs)    = checkExprDuplicates e <|> foldMapAltern
 checkExprDuplicates (Flatten _ e)      = checkExprDuplicates e
 checkExprDuplicates (Annot _ e _)      = checkExprDuplicates e
 checkExprDuplicates Constructor{}      = Nothing
+checkExprDuplicates BuiltinFn{}        = Nothing